requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
WeasyPrint==60.1
celery==5.3.4
redis==5.0.1
htmx==0.1.0
pytest==7.4.2
black==23.9.1
flake8==6.1.0